Update to work with Aluminium GA
This updates pointers to Aluminum versions, adding porting changes.

We disable the odl-l2switch-ui feature, as we do not have dlux
available. It should be pretty easy to add back.

Places where we need to change logic due to Binding mapping having been
changed receive a bit more substantial changes in that:
- we use try-with-resources blocks for read transactions
- dispatch logic is updated to reduce unnecessary nesting
- a few places are also updated to use Uint types

// FIXME: audit code changes

+++ loopremover/implementation/src/main/java/org/opendaylight/l2switch/loopremover/util/InstanceIdentifierUtils.java
@@ -25,6 +25,7 @@ import org.opendaylight.yang.gen.v1.urn.tbd.params.xml.ns.yang.network.topology.
 import org.opendaylight.yang.gen.v1.urn.tbd.params.xml.ns.yang.network.topology.rev131021.network.topology.Topology;
 import org.opendaylight.yang.gen.v1.urn.tbd.params.xml.ns.yang.network.topology.rev131021.network.topology.TopologyKey;
 import org.opendaylight.yangtools.yang.binding.InstanceIdentifier;
+import org.opendaylight.yangtools.yang.common.Uint16;
 
 /**
  * InstanceIdentifierUtils provides utility functions related to InstanceIdentifiers.
@@ -90,7 +91,7 @@ public final class InstanceIdentifierUtils {
      * @return the table id
      */
     public static Short getTableId(final InstanceIdentifier<Table> tablePath) {
-        return tablePath.firstKeyOf(Table.class).getId();
+        return Uint16.valueOf(tablePath.firstKeyOf(Table.class).getId()).shortValue();
     }
 
     /**
